helium: Move helium related files into helium
