+++ zeppelin-web/src/index.js
@@ -58,6 +58,7 @@ import './app/notebook-repository/notebook-repository.controller.js'
 import './app/search/result-list.controller.js'
 import './app/search/search.service.js'
 import './app/helium'
+import './app/helium/helium.service.js'
 import './components/array-ordering/array-ordering.service.js'
 import './components/navbar/navbar.controller.js'
 import './components/navbar/expand-collapse/expand-collapse.directive.js'
@@ -74,4 +75,3 @@ import './components/login/login.controller.js'
 import './components/note-action/note-action.service.js'
 import './components/note-rename/note-rename.controller.js'
 import './components/note-rename/note-rename.service.js'
-import './components/helium/helium.service.js'
